Added back DrawScatter and DrawHeatmap
Copied code over from dviraran/SingleR/master/R/SingleR.Plotting.R, modified DrawScatter and DrawHeatmap code to work with newer data structures, and edited their documentation to reflect the new changes.

Additional notes from Dan:

DrawScatter compares a cell to an individual cell/sample from the reference dataset.  The old method
involved picking an index # from a single table full of all the samples.  In the new SingleR.train structure,
data for cells/samples of the same type are stored separately so the previous indexing method does not work.
My current fix: require provision of the original counts/logcounts dataframe that was used to generate the
training.
Potential alternative: Utilize the trained data, and make the index refer to which of the trained cells of
a given type should be used

DrawHeatmap works with little updating required.  I mainly just changed T/F to TRUE/FALSE here and pointed the
pheatmap calls explicitly to the pheatmap package in case it has not already been loaded into the namespace.

DrawBoxPlot update is in progress