Added sufficient documentation for the min.diff.next cutoff + a bit more
Addressed Aaron's comments

Added myself as an author

Added paragraph in the `details` section describing the goal of the min.diff.next cutoff, it's pitfalls, and its suggested use.

Changed a reference to presumably all scoring measures to be more precise as it needs to not refer to min.diff.next

Removed a `both of` cuz there are now 3, and added min.diff.next to the section about how tweaks to the parameters shoul dbe expected to affect the strength of the cutoffs.