Add support for NA score plotting
Added `na.color` input for setting color of NA scores (=gray50).

Added 'na.rm=TRUE' to score's max/min/range grabs.

In `.plot_score_heatmap` the NAs are replaced by a dummy-value that is colored by 'na.color', and the plot legend is adjusted to call this value "NA".

Added in a direct selection of a color palet for non-normalized data (necessary for the 'na.color' add) and made it color-blindness friendly.

Additional minor bug fix in the `.make_annotation_col`