Update to the latest version of the singlepp library in assorthead. (#274)
This allows us to get rid of our internally vendored copy of the library. 

The latest version of singlepp greatly simplifies management of gene IDs,
forcing us to perform intersections if we are not completely sure that the test
and reference row identities are the same. This is much easier to reason about
than the previous logic; now, our test and reference features are essentially
fixed, and the C++ code handles all of the intersections.

A consequence of this change is that we need to know the test features during
training (or we assume that they're the same as the reference). This involves
some minor changes to the trained output to record the test features, so that
we can check that they're the same as 'rownames(test)' during classification.
If they're not, we throw an error, which is more stringent than the previous
behavior where we would only throw if the subset of markers could not be found;
but in practice, the two are probably the same anyway.

Another consequence is that we can't do any filtering of the test matrix after
training, as that would cause mismatches between the trained output's record of
the test features and the test features used during classificaiton. This is
probably fine as most people shouldn't have NAs in the test matrix, and even
then, most people are using SingleR() and the filtering + intersections are
handled automatically before training anyway.

We also return to using BiocNeighbors. In this case, it provides C++-level
builder objects for the neighbor index construction, allowing us to avoid 
having to compile for all of the different NN search algorithms.