Support alternative row names when plotting the marker heatmap.
We can't ask users to just modify the row names directly, as these are
used to cross-reference the markers in the result metadata. So we
provide a dedicated argument to replace the names, e.g., to switch to
gene symbols instead of IDs.

Also minor fix to the ordering of the labels in the plot. Added an
entry to the NEWS file about this new function.