Cleaned up the per-reference scores from combineRecomputedResults.
Now we return a DataFrame of nested Dataframes, where each inner DataFrame
corresponds to a reference and contains the identity of the best label and the
recomputed score in that reference. This is effectively a sparse version of the
previous "expanded with NA" format, which is easier to read and saves memory.

We also update the plotting functions to work with this new format, mostly by
regenerated the previous expanded format as needed.